refactor : #59 MemberService와 AuthenticationService 분기
